SD2: Move world_eastern_kingdoms to separate file
