+++ src/game/AI/ScriptDevAI/scripts/eastern_kingdoms/world_eastern_kingdoms.cpp
@@ -1,0 +1,60 @@
+/* This file is part of the ScriptDev2 Project. See AUTHORS file for Copyright information
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ */
+
+#include "world_eastern_kingdoms.h"
+#include "AI/ScriptDevAI/include/precompiled.h"
+#include "AI/ScriptDevAI/scripts/world/world_map_scripts.h"
+
+/* *********************************************************
+ *                  EASTERN KINGDOMS
+ */
+struct world_map_eastern_kingdoms : public ScriptedMap
+{
+    world_map_eastern_kingdoms(Map* pMap) : ScriptedMap(pMap) {}
+
+    void OnCreatureCreate(Creature* pCreature) override
+    {
+        switch (pCreature->GetEntry())
+        {
+            case NPC_NEZRAZ:
+            case NPC_HINDENBURG:
+            case NPC_ZAPETTA:
+            case NPC_SQUIBBY_OVERSPECK:
+            case NPC_JONATHAN:
+            case NPC_WRYNN:
+            case NPC_BOLVAR:
+            case NPC_PRESTOR:
+            case NPC_WINDSOR:
+                m_npcEntryGuidStore[pCreature->GetEntry()] = pCreature->GetObjectGuid();
+                break;
+        }
+    }
+
+    void SetData(uint32 /*uiType*/, uint32 /*uiData*/) override {}
+};
+
+InstanceData* GetInstanceData_world_map_eastern_kingdoms(Map* pMap)
+{
+    return new world_map_eastern_kingdoms(pMap);
+}
+
+void AddSC_world_eastern_kingdoms()
+{
+    Script* pNewScript = new Script;
+    pNewScript->Name = "world_map_eastern_kingdoms";
+    pNewScript->GetInstanceData = &GetInstanceData_world_map_eastern_kingdoms;
+    pNewScript->RegisterSelf();
+}
