+++ src/game/AI/ScriptDevAI/scripts/eastern_kingdoms/world_eastern_kingdoms.h
@@ -1,0 +1,15 @@
+
+#ifndef DEF_WORLD_MAP_EK_SCRIPTS_H
+#define DEF_WORLD_MAP_EK_SCRIPTS_H
+
+enum
+{
+    // Quest 6403
+    NPC_JONATHAN                = 466,
+    NPC_WRYNN                   = 1747,
+    NPC_BOLVAR                  = 1748,
+    NPC_PRESTOR                 = 1749,
+    NPC_WINDSOR                 = 12580,
+};
+
+#endif
