+++ src/game/AI/ScriptDevAI/scripts/world/world_map_scripts.cpp
@@ -29,39 +29,6 @@ EndScriptData */
 #include <array>
 #include <ctime>
 
-/* *********************************************************
- *                  EASTERN KINGDOMS
- */
-struct world_map_eastern_kingdoms : public ScriptedMap
-{
-    world_map_eastern_kingdoms(Map* pMap) : ScriptedMap(pMap) {}
-
-    void OnCreatureCreate(Creature* pCreature)
-    {
-        switch (pCreature->GetEntry())
-        {
-            case NPC_NEZRAZ:
-            case NPC_HINDENBURG:
-            case NPC_ZAPETTA:
-            case NPC_SQUIBBY_OVERSPECK:
-            case NPC_JONATHAN:
-            case NPC_WRYNN:
-            case NPC_BOLVAR:
-            case NPC_PRESTOR:
-            case NPC_WINDSOR:
-                m_npcEntryGuidStore[pCreature->GetEntry()] = pCreature->GetObjectGuid();
-                break;
-        }
-    }
-
-    void SetData(uint32 /*uiType*/, uint32 /*uiData*/) {}
-};
-
-InstanceData* GetInstanceData_world_map_eastern_kingdoms(Map* pMap)
-{
-    return new world_map_eastern_kingdoms(pMap);
-}
-
 enum
 {
     EVENT_UC_FROM_GROMGOL_ARRIVAL = 15312,
@@ -111,11 +78,6 @@ bool ProcessEventTransports(uint32 uiEventId, Object* pSource, Object* /*pTarget
 void AddSC_world_map_scripts()
 {
     Script* pNewScript = new Script;
-    pNewScript->Name = "world_map_eastern_kingdoms";
-    pNewScript->GetInstanceData = &GetInstanceData_world_map_eastern_kingdoms;
-    pNewScript->RegisterSelf();
-
-    pNewScript = new Script;
     pNewScript->Name = "event_transports";
     pNewScript->pProcessEventId = &ProcessEventTransports;
     pNewScript->RegisterSelf();
