+++ src/game/AI/ScriptDevAI/scripts/world/world_map_scripts.h
@@ -7,13 +7,6 @@
 
 enum
 {
-    // Quest 6403
-    NPC_JONATHAN                = 466,
-    NPC_WRYNN                   = 1747,
-    NPC_BOLVAR                  = 1748,
-    NPC_PRESTOR                 = 1749,
-    NPC_WINDSOR                 = 12580,
-
     // Transports
     NPC_NEZRAZ                  = 3149,
     NPC_HINDENBURG              = 3150,
